academy/requirements.txt
amqp==5.0.5
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-bulk-update-or-create==0.2.0
django-factory-boy==1.0.0
django-silk==4.1.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.0.0
gprof2dot==2019.11.30
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pipdeptree==2.0.0
pluggy==0.13.1
prompt-toolkit==3.0.16
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dateutil==2.8.1
python-http-client==3.3.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
sendgrid==6.6.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
